Add files via upload
A **Decision Tree** is a supervised learning model used for classification and regression. It splits data into branches based on feature values, forming a tree-like structure. You provide input and output data, train the model on 25% of it, and adjust it to minimize errors. It predicts outcomes based on learned patterns.